+++ smart_contracts/vm2/sdk/Cargo.toml
@@ -26,6 +26,7 @@ num-traits = "0.2.19"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 thiserror = { version = "2", optional = true }
+hashbrown = { version = "0.16", optional = true }
 
 [dev-dependencies]
 casper-types = { path = "../../../types", features = ["std", "testing", "gens"] }
