+++ smart_contracts/vm2/sdk/src/abi_generator.rs
@@ -1,4 +1,4 @@
-use core::{mem, ptr::NonNull};
+use crate::prelude::{mem, ptr::NonNull, *};
 
 use crate::{
     abi::{Declaration, Definitions},
