+++ smart_contracts/vm2/sdk/src/collections/set.rs
@@ -1,18 +1,11 @@
-use crate::{
-    abi::{CasperABI, Declaration, Definition, Definitions, StructField},
-    prelude::marker::PhantomData,
-};
+use crate::prelude::marker::PhantomData;
 
-use crate::{
-    casper,
-    serializers::borsh::{BorshDeserialize, BorshSerialize},
-};
-use casper_executor_wasm_common::{error::HostResult, keyspace::Keyspace};
+use crate::{casper, prelude::*, serializers::borsh::BorshSerialize};
+use casper_executor_wasm_common::keyspace::Keyspace;
 
 use super::lookup_key::{Identity, LookupKey, LookupKeyOwned};
 
-#[derive(BorshSerialize, BorshDeserialize, Debug, Clone)]
-#[borsh(crate = "crate::serializers::borsh")]
+#[derive(Clone)]
 pub struct Set<T, L = Identity>
 where
     T: BorshSerialize,
@@ -20,29 +13,6 @@ where
     prefix: String,
     lookup: L,
     _marker: PhantomData<T>,
-}
-
-impl<T: CasperABI + BorshSerialize, L> CasperABI for Set<T, L> {
-    fn populate_definitions(_definitions: &mut Definitions) {}
-
-    fn declaration() -> Declaration {
-        format!("Set<{}>", T::declaration())
-    }
-
-    fn definition() -> Definition {
-        Definition::Struct {
-            items: vec![
-                StructField {
-                    name: "prefix".into(),
-                    decl: String::declaration(),
-                },
-                StructField {
-                    name: "length".into(),
-                    decl: u64::declaration(),
-                },
-            ],
-        }
-    }
 }
 
 impl<T, L> Set<T, L>
@@ -64,25 +34,16 @@ where
         casper::write(Keyspace::Context(lookup_key.as_ref()), &[]).unwrap();
     }
 
-    pub fn contains(&self, key: &T) -> bool {
-        let lookup_key = self.lookup.lookup(self.prefix.as_bytes(), key);
+    pub fn contains_key(&self, key: T) -> bool {
+        let lookup_key = self.lookup.lookup(self.prefix.as_bytes(), &key);
         let entry = casper::read(Keyspace::Context(lookup_key.as_ref()), |_size| None).unwrap();
         entry.is_some()
-    }
-
-    pub fn remove(&mut self, key: &T) -> bool {
-        let lookup_key = self.lookup.lookup(self.prefix.as_bytes(), key);
-        match casper::remove(Keyspace::Context(lookup_key.as_ref())) {
-            Ok(()) => true,
-            Err(HostResult::NotFound) => false,
-            Err(other) => panic!("Error removing from set: {:?}", other),
-        }
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::{casper::native::dispatch, prelude::*};
+    use crate::prelude::*;
 
     use crate::serializers::borsh::BorshSerialize;
 
@@ -96,24 +57,22 @@ mod tests {
         C,
     }
 
+    #[ignore]
     #[test]
     fn should_insert() {
-        dispatch(|| {
-            let mut set: Set<Flag> = Set::new("Prefix".to_string());
+        let mut set: Set<Flag> = Set::new("Prefix".to_string());
 
-            assert!(!set.contains(&Flag::A));
-            assert!(!set.contains(&Flag::B));
-            assert!(!set.contains(&Flag::C));
+        assert!(!set.contains_key(Flag::A));
+        assert!(!set.contains_key(Flag::B));
+        assert!(!set.contains_key(Flag::C));
 
-            set.insert(Flag::A);
-            assert!(set.contains(&Flag::A));
+        set.insert(Flag::A);
+        assert!(set.contains_key(Flag::A));
 
-            set.insert(Flag::B);
-            assert!(set.contains(&Flag::B));
+        set.insert(Flag::B);
+        assert!(set.contains_key(Flag::B));
 
-            set.insert(Flag::C);
-            assert!(set.contains(&Flag::C));
-        })
-        .unwrap();
+        set.insert(Flag::C);
+        assert!(set.contains_key(Flag::C));
     }
 }
