+++ smart_contracts/vm2/sdk/src/contrib/pausable.rs
@@ -10,7 +10,7 @@
 //!
 //! For security reasons you may want to combine `AccessControl` or `Ownable` with
 //! this trait to ensure that only selected entities can manage the pause state.
-use crate::{self as casper_contract_sdk, casper, macros::casper, types::entity::Entity};
+use crate::{self as casper_contract_sdk, casper, casper::Entity};
 
 #[casper]
 pub struct PausedState {
